pyproject.toml
[project]
name = "storb"
version = "0.1.0"
description = "An object storage subnet on the Bittensor network"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.20.0",
    "colorlog>=6.9.0",
    "cryptography>=44.0.0",
    "fastapi>=0.112.0",
    "fiber @ git+https://github.com/fr34kcoders/fiber.git",
    "httpx>=0.28.1",
    "kademlia>=2.2.2",
    "numpy==2.0.2",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.19",
    "rich>=13.9.4",
    "starlette==0.37.2",
    "substrate-interface>=1.7.10",
    "uvicorn==0.30.5",
    "zfec>=1.6.0.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.9",
    "py-spy>=0.4.0",
    "pytest>=8.3.4",
    "ruff>=0.8.2",
]

[tool.ruff]
exclude = [
    ".git",
    ".pyenv",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "venv",
]
indent-width = 4

[tool.ruff.lint]
fixable = ["ALL"]
ignore = ["E501"]
select = ["F", "E", "W", "N", "I", "A", "PLC", "PLE", "PLW"]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
quote-style = "double"

[tool.setuptools]
packages = ["storb"]

[tool.setuptools.package-dir]
storb = "storb"
